move ArmorConfig into seperate class + gradle changes
