+++ gradle.properties
@@ -1,6 +1,8 @@
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
+org.gradle.warning.mode=all
+org.gradle.caching=true
 
 # Fabric Properties
 	# check these on https://fabricmc.net/develop
