+++ src/main/java/io/github/offbeat_stuff/zombie_apocalypse/mixin/ZombieEntityMixin.java
@@ -1,8 +1,8 @@
 package io.github.offbeat_stuff.zombie_apocalypse.mixin;
 
+import io.github.offbeat_stuff.zombie_apocalypse.ArmorHandler;
 import io.github.offbeat_stuff.zombie_apocalypse.ZombieEntityInterface;
 import io.github.offbeat_stuff.zombie_apocalypse.ZombieKind;
-import io.github.offbeat_stuff.zombie_apocalypse.config.ArmorHandler;
 import io.github.offbeat_stuff.zombie_apocalypse.config.ConfigHandler;
 import io.github.offbeat_stuff.zombie_apocalypse.config.WeaponHandler;
 import net.minecraft.entity.Entity;
@@ -11,7 +11,6 @@ import net.minecraft.server.world.ServerWorld;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-// import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(ZombieEntity.class)
